+++ src/app/(connected)/_components/app-command-dialog.tsx
@@ -119,7 +119,7 @@ export function AppCommandDialog() {
 
   return (
     <Dialog open={isOpen} onOpenChange={handleChangeDialogState}>
-      <DialogContent className="inset-0 max-h-none overflow-hidden rounded-none p-0 duration-500 data-[state=closed]:zoom-out-100 data-[state=open]:zoom-in-100 data-[state=closed]:slide-out-to-bottom-0 data-[state=closed]:slide-out-to-right data-[state=open]:slide-in-from-bottom-0 data-[state=open]:slide-in-from-right sm:right-auto sm:top-auto sm:h-[90vh] sm:max-h-96 sm:max-w-2xl sm:p-0">
+      <DialogContent className="inset-0 max-h-none overflow-hidden rounded-none border-0 p-0 duration-500 data-[state=closed]:zoom-out-100 data-[state=open]:zoom-in-100 data-[state=closed]:slide-out-to-bottom-0 data-[state=closed]:slide-out-to-right data-[state=open]:slide-in-from-bottom-0 data-[state=open]:slide-in-from-right sm:right-auto sm:top-auto sm:h-[90vh] sm:max-h-96 sm:max-w-2xl sm:border sm:p-0">
         <Command
           shouldFilter={false}
           loop
@@ -371,7 +371,7 @@ export function AppCommandDialogTrigger(props: AppCommandDialogTriggerProps) {
         >
           <SearchIcon className="size-6 shrink-0" />
           <span className="sr-only">Search</span>
-          <span className="hidden rounded-sm bg-surface-hover p-1 text-base-s-regular text-muted-foreground sm:inline-block">
+          <span className="hidden rounded-sm bg-surface-hover p-1 text-base-s-regular text-muted-foreground md:inline-block">
             {shortcutText}
           </span>
         </Button>