+++ src/app/(connected)/connections/summary/logs/layout.tsx
@@ -1,5 +1,9 @@
 import { notFound } from "next/navigation"
 
+import { DestroyAllLogsDialog } from "@/app/(connected)/connections/summary/logs/_components/destroy-all-logs-dialog"
+import { DeleteIcon } from "@/components/icons/delete-icon"
+import { PageWrapper } from "@/components/page-wrapper"
+import { Button } from "@/components/ui/button"
 import { featureFlags } from "@/config/features"
 
 type ConnectionsLogsLayoutProps = {
@@ -15,6 +19,25 @@ export default function ConnectionsLogsLayout(
     notFound()
   }
 
-  return children
+  return (
+    <PageWrapper
+      pageTitle="Connection Logs"
+      rightContent={
+        featureFlags.dataConnections.logs.destroy ? (
+          <DestroyAllLogsDialog>
+            <Button
+              variant="outline-destructive"
+              className="h-12 w-12 p-0 sm:w-auto sm:px-6 sm:py-2"
+            >
+              <DeleteIcon className="size-5 shrink-0 sm:hidden" />
+              <span className="sr-only sm:not-sr-only">Delete All</span>
+            </Button>
+          </DestroyAllLogsDialog>
+        ) : undefined
+      }
+    >
+      {children}
+    </PageWrapper>
+  )
 }
 ConnectionsLogsLayout.displayName = "ConnectionsLogsLayout"