+++ src/app/(connected)/data/layout.tsx
@@ -1,6 +1,11 @@
+import { Metadata } from "next"
 import { notFound } from "next/navigation"
 
 import { featureFlags } from "@/config/features"
+
+export const metadata: Metadata = {
+  title: "Data",
+}
 
 type DataLayoutProps = {
   children: React.ReactNode